requirements.txt
pandas>=2.0
pyarrow>=14.0
openpyxl>=3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyYAML>=6.0
gigachat
sqlparse>=0.4.0
pytest>=7.0.0
mcp>=1.0.0
gradio
fastapi
uvicorn[standard]
pydantic>=2.0
markdown>=3.4
weasyprint>=60.0